+++ src/MoviesService.java
@@ -1,7 +1,9 @@
+import java.math.BigDecimal;
+
 public interface MoviesService {
 
   public Movie getMovieById(String movieId) throws Exception;
 
-  public double calculateRentPrice(MovieCode movieCode, int days) throws Exception;
+  public BigDecimal calculateRentPrice(MovieCode movieCode, int days) throws Exception;
 
 }
